requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pretty-errors==1.2.25
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==1.3.1
setuptools==69.5.1
six==1.16.0
tweepy==4.14.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.43.0
